:art: Move filter registration to CLI and rename ExampleFilter to DoublingFilter
