+++ ice_station_zebra/config/datasets/nbsamp_weathersouth_era5_0p5_2020_2020_24h_v1.yaml
@@ -11,7 +11,11 @@ nbsamp-weathersouth-era5-0p5-2020-2020-24h-v1:
     frequency: 24h
 
   input:
-    anemoi-dataset:
-      join:
-      - dataset: ${base_path}/data/anemoi/samp-weathersouth-era5-0p5-2020-2020-24h-v1.zarr
-        select: [ 2t, z_250, z_500 , 10u, 10v, cos_julian_day, sin_julian_day ]
+    pipe:
+      - anemoi-dataset:
+          join:
+          - dataset: ${base_path}/data/anemoi/samp-weathersouth-era5-0p5-2020-2020-24h-v1.zarr
+            select: [ 2t, z_250, z_500 , 10u, 10v, cos_julian_day, sin_julian_day ]
+      - doubling_filter:
+          input_field: 2t
+          output_field: doubled_2t
