+++ ice_station_zebra/data_processors/zebra_data_processor.py
@@ -4,13 +4,11 @@ from pathlib import Path
 
 from anemoi.datasets.commands.create import Create
 from anemoi.datasets.commands.inspect import InspectZarr
-from anemoi.transform.filters import filter_registry
 from omegaconf import DictConfig, OmegaConf
 from zarr.errors import PathNotFoundError
 
 from ice_station_zebra.types import AnemoiCreateArgs, AnemoiInspectArgs
 
-from .filters.example_filter import ExampleFilter
 from .preprocessors import IPreprocessor
 
 logger = logging.getLogger(__name__)
@@ -48,7 +46,6 @@ class ZebraDataProcessor:
 
     def download(self) -> None:
         """Download a single Anemoi dataset."""
-        filter_registry.register("example_filter", ExampleFilter)
         self.preprocessor.download(self.path_preprocessor)
         logger.info("Creating dataset %s at %s.", self.name, self.path_dataset)
         Create().run(
